Move handleSwipe definition to where it's used
