+++ assets/src/components/propertiesPanel.tsx
@@ -1,18 +1,29 @@
 import React, { useContext, useEffect } from "react"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
-import detectSwipe from "../helpers/detectSwipe"
+import detectSwipe, { SwipeDirection } from "../helpers/detectSwipe"
 import { isAVehicle } from "../models/vehicle"
 import { VehicleOrGhost } from "../realtime.d"
 import { Route } from "../schedule"
 import { deselectVehicle } from "../state"
 import GhostPropertiesPanel from "./propertiesPanel/ghostPropertiesPanel"
-import VehiclePropertiesPanel, {
-  handleSwipe,
-} from "./propertiesPanel/vehiclePropertiesPanel"
+import VehiclePropertiesPanel from "./propertiesPanel/vehiclePropertiesPanel"
 
 interface Props {
   selectedVehicleOrGhost: VehicleOrGhost
   route?: Route
+}
+
+export const handleSwipe = (hideMe: () => void) => (
+  swipeDirection: SwipeDirection,
+  target: HTMLElement | null
+) => {
+  if (target && target.id === "id-vehicle-map") {
+    return
+  }
+
+  if (swipeDirection === "Right") {
+    hideMe()
+  }
 }
 
 const PropertiesPanel = ({ selectedVehicleOrGhost, route }: Props) => {
