+++ assets/tests/components/propertiesPanel/vehiclePropertiesPanel.test.tsx
@@ -1,9 +1,7 @@
 import { mount } from "enzyme"
 import React from "react"
 import renderer from "react-test-renderer"
-import VehiclePropertiesPanel, {
-  handleSwipe,
-} from "../../../src/components/propertiesPanel/vehiclePropertiesPanel"
+import VehiclePropertiesPanel from "../../../src/components/propertiesPanel/vehiclePropertiesPanel"
 import { StateDispatchProvider } from "../../../src/contexts/stateDispatchContext"
 import { HeadwaySpacing } from "../../../src/models/vehicleStatus"
 import { Vehicle } from "../../../src/realtime"
@@ -199,31 +197,5 @@ describe("VehiclePropertiesPanel", () => {
     wrapper.find(".m-properties-panel__close-button").simulate("click")
 
     expect(mockDispatch).toHaveBeenCalledWith(deselectVehicle())
-  })
-})
-
-describe("handleSwipe", () => {
-  test("hides the panel on a right swipe", () => {
-    const hidePanelCB = jest.fn()
-
-    handleSwipe(hidePanelCB)("Right", null)
-    expect(hidePanelCB).toHaveBeenCalled()
-  })
-
-  test("does not hide panel on other swipes", () => {
-    const hidePanelCB = jest.fn()
-    const handler = handleSwipe(hidePanelCB)
-    handler("Left", null)
-    handler("Up", null)
-    handler("Down", null)
-
-    expect(hidePanelCB).not.toHaveBeenCalled()
-  })
-
-  test("does not hide panel when map is swiped right", () => {
-    const hidePanelCB = jest.fn()
-    const map = document.createElement("div")
-    map.id = "id-vehicle-map"
-    handleSwipe(hidePanelCB)("Right", map)
   })
 })
